requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==7.1.2
cycler==0.10.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.24
idna==3.2
imageio==2.9.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
ptyprocess==0.7.0
pyarrow==5.0.0
pycparser==2.20
pydeck==0.7.0
pydicom==2.2.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
pyzmq==22.3.0
requests==2.26.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
Send2Trash==1.8.0
simplejson==3.17.5
six==1.16.0
sklearn==0.0
smmap==4.0.0
streamlit==1.1.0
streamlit-aggrid==0.2.2.post2
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.10.12
toml==0.10.2
toolz==0.11.1
torch==1.9.1
torchvision==0.10.1
torchxrayvision==0.0.32
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
